Move TaskScheduler and friends into mzutils
