+++ src/addons/manager/addonmanager.cpp
@@ -27,7 +27,7 @@
 #include "settingsholder.h"
 #include "tasks/addon/taskaddon.h"
 #include "tasks/addonindex/taskaddonindex.h"
-#include "tasks/function/taskfunction.h"
+#include "taskfunction.h"
 #include "taskscheduler.h"
 
 constexpr const char* MVPN_ENV_SKIP_ADDON_SIGNATURE =
