+++ src/controller.cpp
@@ -32,7 +32,7 @@
 #include "serverlatency.h"
 #include "settingsholder.h"
 #include "tasks/controlleraction/taskcontrolleraction.h"
-#include "tasks/function/taskfunction.h"
+#include "taskfunction.h"
 #include "taskscheduler.h"
 
 #if defined(MZ_FLATPAK)
