+++ src/mozillavpn.cpp
@@ -51,13 +51,13 @@
 #include "tasks/controlleraction/taskcontrolleraction.h"
 #include "tasks/createsupportticket/taskcreatesupportticket.h"
 #include "tasks/deleteaccount/taskdeleteaccount.h"
-#include "tasks/function/taskfunction.h"
 #include "tasks/getlocation/taskgetlocation.h"
 #include "tasks/getsubscriptiondetails/taskgetsubscriptiondetails.h"
-#include "tasks/group/taskgroup.h"
 #include "tasks/heartbeat/taskheartbeat.h"
 #include "tasks/removedevice/taskremovedevice.h"
 #include "tasks/servers/taskservers.h"
+#include "taskfunction.h"
+#include "taskgroup.h"
 #include "taskscheduler.h"
 #include "telemetry.h"
 #include "update/updater.h"
